chore(test): refactor request executors
